refactor: re-organizando diretorios modulo base interfaces
