+++ src/modules/base/interfaces/domain/base.squema.interface.ts
