+++ src/modules/base/interfaces/repository/delete.generic.base.interface.ts
