+++ src/modules/base/interfaces/repository/deleteMany.generic.base.interface.ts
