+++ src/modules/base/interfaces/repository/get-all.generic.base.interface.ts
