+++ src/modules/base/interfaces/repository/get-one.generic.base.interface.ts
